+++ zcash_client_backend/src/data_api.rs
@@ -1,7 +1,7 @@
 //! Interfaces for wallet data persistence & low-level wallet utilities.
 
 use std::cmp;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::fmt::Debug;
 
 use secrecy::SecretVec;
@@ -17,7 +17,7 @@ use zcash_primitives::{
 };
 
 use crate::{
-    address::UnifiedAddress,
+    address::{AddressMetadata, UnifiedAddress},
     decrypt::DecryptedOutput,
     keys::{UnifiedFullViewingKey, UnifiedSpendingKey},
     proto::compact_formats::CompactBlock,
@@ -207,7 +207,7 @@ pub trait WalletRead {
     fn get_transparent_receivers(
         &self,
         account: AccountId,
-    ) -> Result<HashSet<TransparentAddress>, Self::Error>;
+    ) -> Result<HashMap<TransparentAddress, AddressMetadata>, Self::Error>;
 
     /// Returns a list of unspent transparent UTXOs that appear in the chain at heights up to and
     /// including `max_height`.
@@ -400,9 +400,6 @@ pub mod testing {
     use secrecy::{ExposeSecret, SecretVec};
     use std::collections::HashMap;
 
-    #[cfg(feature = "transparent-inputs")]
-    use std::collections::HashSet;
-
     use zcash_primitives::{
         block::BlockHash,
         consensus::{BlockHeight, Network},
@@ -415,7 +412,7 @@ pub mod testing {
     };
 
     use crate::{
-        address::UnifiedAddress,
+        address::{AddressMetadata, UnifiedAddress},
         keys::{UnifiedFullViewingKey, UnifiedSpendingKey},
         proto::compact_formats::CompactBlock,
         wallet::{SpendableNote, WalletTransparentOutput},
@@ -555,8 +552,8 @@ pub mod testing {
         fn get_transparent_receivers(
             &self,
             _account: AccountId,
-        ) -> Result<HashSet<TransparentAddress>, Self::Error> {
-            Ok(HashSet::new())
+        ) -> Result<HashMap<TransparentAddress, AddressMetadata>, Self::Error> {
+            Ok(HashMap::new())
         }
 
         fn get_unspent_transparent_outputs(